RTC and Alarm working
Got the RTC and alarm working. Simplified the structure of the RTOS as previous
implementation was quite cumbersome. timeKeeping task is not in use at the moment.
The software now displays current time and upon moisture detection (SW2) sets up
alarm for 10s after the current time (for testing purposes).