Original functionality restored
I am back to original functionality after acquiring
the J-Link. As said before I had to start using SEGGER's own
RTTViewer instead of my debug console. This leads to a fact
that I am currently unable to use UART0 and it's interrupts as
it is not connected to secondary USB on K64F. I have worked around
it by creating another task that checks the RTT input periodically
and responds to it. This should be enough for my test and debug
purposes as eventually this system should be replaced by an LCD and
a dedicated keypad.