timeConfig working
time configuration task is working. Thanks to my work on another mini
project I was able to reuse some code written by me in this one.
userTimeConfig task requires a decent amount of memory which was one
of the problems I haven't considered previously.
UART interrupt is disabled for the moment but will attempt to get it
functional again. However, UART0 isn't too important for the project as
eventually user input and output should be delivered via LCD display.