bluetooth pins configured
configured pins for communication with bluetooth module, however it is
a bit confusing. According to schematic they should connect to
PTC14 and PTC15 however these are not on the headers. Tried to enable
UART4 functionality on PTE24 and PTE25 as the pins alternative function
but I am not getting any response from the BT module so far.