bluetooth receiving
bluetooth finally receiving and printf is working (almost flawlessly).
It seems that I have been using the UART4 interrupts in wrong way. At
the moment I am using just a simple polling method and everything seems
to work. I have created a new task btTask. Upon receiving a character
from my phone in mainTask a btTask is created and just prints the
character at the moment. This isn't the meant functionality as the Dev1
will be mainly sending notifications to Dev2 to start alarm but could be
used to get information about the state of the Dev2 and print it out.