communication from phone to Dev2 established
Queue for transmission of data between phoneBTTask and bluetooth task
created. I can now send a character 'a' from the phone that will get
transferred from phoneBTTask via queue to bluetooth task that will in
turn send this via bluetooth to Dev2. Delighted this is working.
Noticed issues with printing in the terminal. Will have to implement
safe printf as I think too many resources are trying to print at the
same time